Command registration: Add default permissions when not found
If you don't add the `defaultPermission` property to a command before registering it, Discord will automatically add `defaultPermission: true` to that command's data. This was causing Hiei's automatic command registration to always be out of sync. In this update, Hiei now adds `defaultPermission: true` before registering if the property isn't present.