Inject required parameter before syncing command data
When commands are registered with a guild, optional arguments have `required: false` set on them. This was causing command sync to break because I only add the `required` parameter if it's true. To fix this, I now inject `required: false` for optional arguments before comparing local data with remote data.